// src/bioma/domain/types/UserInterface.ts
export interface UserInterface {
    id_user: number;
    user_name: string;
    email: string,
    password: string;
    name: string, 
    pet_name: string;
    education: string, 
    birthdate: Date,
    telephone:string, 
    sex: string, 
    occupation: string,
    time_played: number;
    questions_learned: number;
    received_xp: number;
    streak: string;
    last_date_added: Date;
    league: string; 
}